mv push down filter entirely to core tests
