+++ datafusion/core/tests/config_from_env.rs
@@ -15,6 +15,8 @@
 // specific language governing permissions and limitations
 // under the License.
 
+mod optimizer;
+
 use datafusion::config::ConfigOptions;
 use std::env;
